+++ modules/vt.py
@@ -14,7 +14,7 @@ from pyrogram import Client, filters, enums
 from pyrogram.types import Message
 
 from utils.misc import modules_help, prefix
-from utils.scripts import edit_or_reply, format_exc, time_formatter, humanbytes, edit_or_send_as_file, get_text, progress
+from utils.scripts import edit_or_reply, format_exc, progress
 from utils.config import vt_key as vak
 
 
@@ -48,7 +48,8 @@ async def scan_my_file(client, message):
     except Exception as e:
       return await ms_.edit(format_exc(e))
     await ms_.edit(f"<b><u>Scanned {message.reply_to_message.document.file_name}</b></u>. <b>You Can Visit :</b> <a href=\"https://www.virustotal.com/gui/file/{md5}\">Here</a> <b>In 5-10 Min To See File Report</b>")
-    os.remove(downloaded_file_name)
+    if os.path.exists(downloaded_file_name):
+      os.remove(downloaded_file_name)
 
 @Client.on_message(filters.command("vtl", prefix) & filters.me)
 async def scan_my_large_file(client, message):
@@ -119,7 +120,8 @@ async def scan_my_large_file(client, message):
     except Exception as e:
       return await ms_.edit(format_exc(e))
     await ms_.edit(f"<b><u>Scanned {message.reply_to_message.document.file_name}</b></u>. <b>You Can Visit :</b> <a href=\"https://www.virustotal.com/gui/file/{md5}\">Here</a> <b>In 5-10 Min To See File Report</b>")
-    os.remove(downloaded_file_name)
+    if os.path.exists(downloaded_file_name):
+      os.remove(downloaded_file_name)
 
 modules_help["virustotal"] = {
     "vt [reply to file]*": "Scan for viruses on Virus Total (for lower file size <32MB)",
